+++ src/index.ts
@@ -1,38 +1,20 @@
-import express, { RequestHandler } from 'express';
+import express from 'express';
 import connectDB from './loaders/db';
 import routes from './routes';
 import helmet from 'helmet';
 import errorHandler from './middlewares/errorHandler';
 import * as SentryConfig from './loaders/sentryConfiguration';
 import corsMiddleware from './middlewares/cors';
-import session from 'express-session';
-import MongoStore from 'connect-mongo';
 import config from './config';
-
-const expressSession: RequestHandler = session({
-  secret: config.sessionKey,
-  store: MongoStore.create({
-    mongoUrl: config.mongoURI
-  }),
-  cookie: {
-    sameSite: 'none',
-    httpOnly: true,
-    secure: true,
-    maxAge: 3.6e6 * 24 * 180
-  },
-  resave: false,
-  proxy: true,
-  saveUninitialized: false
-});
+import expressSession from './middlewares/session/guestSession';
 
 const app = express();
 app.use(express.urlencoded({ extended: true }));
 app.use(express.json());
 app.use(helmet());
-SentryConfig.initializeSentry(app);
-
-app.use(expressSession);
 app.use(corsMiddleware);
+app.use(expressSession);
+SentryConfig.initializeSentry(app);
 app.use(routes);
 SentryConfig.attachSentryErrorHandler(app);
 app.use(errorHandler);
