+++ src/middlewares/session/guestSession.ts
@@ -1,15 +1,21 @@
-import express from 'express';
 import session from 'express-session';
-import config from '../config';
 import MongoStore from 'connect-mongo';
+import config from '../../config';
+import { RequestHandler } from 'express';
 
-const expressSession: express.RequestHandler = session({
+const expressSession: RequestHandler = session({
   secret: config.sessionKey,
   store: MongoStore.create({
     mongoUrl: config.mongoURI
   }),
-  cookie: { maxAge: 3.6e6 * 24 * 180 },
+  cookie: {
+    sameSite: 'none',
+    httpOnly: true,
+    secure: true,
+    maxAge: 3.6e6 * 24 * 180
+  },
   resave: false,
+  proxy: true,
   saveUninitialized: false
 });
